+++ MetaData/MetaDataAwareInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bugsnag\BugsnagBundle\MetaData;
+
+interface MetaDataAwareInterface
+{
+    /**
+     * Get meta data.
+     *
+     * @return array
+     */
+    public function getMetaData();
+}